lavf: do not use int to store an int64

+++ libavformat/utils.c
@@ -77,7 +77,7 @@ const char *avformat_license(void)
 static int append_packet_chunked(AVIOContext *s, AVPacket *pkt, int size)
 {
     int64_t chunk_size = size;
-    int orig_pos       = pkt->pos; // av_grow_packet might reset pos
+    int64_t orig_pos   = pkt->pos; // av_grow_packet might reset pos
     int orig_size      = pkt->size;
     int ret = 0;
